% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataFiles.R
\docType{data}
\name{CandySet}
\alias{CandySet}
\title{Candy Set}
\format{An object of class \code{data.frame} with 1842 rows and 4 columns.}
\usage{
CandySet
}
\description{
In this dataset, there are 6 kinds of candies for people to choose from.
Participants rated how much they liked each of the 6 candies using an
11-point Likert scale (1 = "dislike very much" and 11 = "like very much"),
which gives us the covariate "Rate". Following that, participants ranked the
6 candies, from their most favorite (rank = 1) to their least favorite (rank
= 12), which accounts for the covariate "Rank".
}
\keyword{datasets}
