% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataFiles.R
\docType{data}
\name{CFS5_1}
\alias{CFS5_1}
\title{Consideration Set Simulation 1}
\format{An object of class \code{data.frame} with 25000 rows and 4 columns.}
\usage{
CFS5_1
}
\description{
NumOpts=5;
NumPeople=5000;
NumCovs=3;
BetaConsid=[2 0 -1];
BetaSelect=[0 -.5 1];
ASCsConsid=[1 2 3 4 5]/5;
ASCsSelect=-[4 3 2 1 0]/4; % Last to zero for identification
}
\keyword{datasets}
