% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataFiles.R
\docType{data}
\name{CFS5_2}
\alias{CFS5_2}
\title{Consideration Set Simulation 3}
\format{An object of class \code{data.frame} with 25000 rows and 5 columns.}
\usage{
CFS5_2
}
\description{
Here we generate some data where (1) there are "multiplicities" (an outcome
choice vector can be like [2 0 1 0 0]); and (2) some items are "not
available". We use the same values as before, but with 10% of the items not
available, and an average number of chosen items = 1.8 (instead of 1).
}
\keyword{datasets}
