% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePoint.R
\name{change.normal}
\alias{change.normal}
\title{Changepoint Model with Normal Distribution}
\usage{
change.normal(v, pre = NA, type = "MeanVar", know.mean = FALSE, mu = NA)
}
\arguments{
\item{v}{Vector that contains response variable of interest.}

\item{pre}{A value  represents preselected changepoint for any model, should
the user wish to test a specific point in data; NA by default.}

\item{type}{Choose between "Mean", "Var", and "MeanVar", regarding the
parameter(s) for which you wish to test significant change in value;
"MeanVar" by default.}

\item{know.mean}{If \code{type="Var"}, Boolean variable indicating whether or
not you wish to set value for "true mean" of data; FALSE by default.}

\item{mu}{If \code{know.mean = TRUE}, single value that represents "true
mean" of data; NA by default.}
}
\value{
List of the following: Table that contains location of
  changepoint(s) (either preselected or determined rigorously) ranked by
  significance and its/their associated significance value(s); Single value
  representing the MBIC threshold by which to base significance of a
  changepoint (i.e., penalty values LARGER than MBIC threshold indicate
  significance)
}
\description{
Fit changepoint model on data to test if there is a significant change in
mean and/or variance of data.
}
\examples{
a = rnorm(100,0,25/100)
b = rep(0,100)
for (i in 1:100) {
  b[i] = rnorm(1,0,(i/2+25)/100)
}
c = rnorm(100,0,75/100)
d = rep(0,100)
for (i in 1:100)
  d[i] = rnorm(1,0,(75-i/2)/100)
e = rnorm(100,0,25/100)
v = c(a,b,c,d,e)

change.normal(v, type="Var", know.mean=TRUE, mu=0)
change.normal(v, type="Var", know.mean=TRUE, mu=0, pre=200)

}
