% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePoint.R
\name{change.poisson}
\alias{change.poisson}
\title{Changepoint Model with Poisson Distribution}
\usage{
change.poisson(v, pre = NA)
}
\arguments{
\item{v}{Vector that contains response variable of interest.}

\item{pre}{A value represents preselected changepoint for any model, should
the user wish to test a specific point in data, NA by default.}
}
\value{
List of the following: Table that contains location of
  changepoint(s) (either preselected or determined rigorously) ranked by
  significance and its/their associated significance value(s); Single value
  representing the MBIC threshold by which to base significance of a
  changepoint (i.e., penalty values LARGER than MBIC threshold indicate
  significance)
}
\description{
Fit changepoint model on data to test if there is a significant change in
Poisson rate parameter of data.
}
\examples{
#Examples you wish to use, in demonstration of function
a = rpois(200,1)
b = rpois(200,2)
c = rpois(200,3)
v = c(a,b,c)

change.poisson(v)
change.poisson(v, pre=400)

}
