% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataFiles.R
\docType{data}
\name{dating}
\alias{dating}
\title{Dating}
\format{An object of class \code{data.frame} with 405249 rows and 4 columns.}
\usage{
dating
}
\description{
Response variables are choices that apply to different stages. The first
stage corresponds to whether or not a user browsed a profile. The second
stage corresponds to whether or not a user wrote to a profile. For each
stage, we have only one identical covariate "agedif".
}
\keyword{datasets}
