% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCmods-package.r
\docType{package}
\name{DCmods}
\alias{DCmods}
\alias{DCmods-package}
\title{DCmods.}
\description{
This package is a collection of various statistical models that can prove to
be useful for marketing and social research on choice behavior data. The
models included are the changepoint model, consideration set model, and
latent stage model.
}
\details{
For details on each model, please refer to the package's vignette, "Better
Understanding DCmods," as well as the help pages of each function.
}
\author{
Elizabeth Bruch, Fred Feinberg, Cameron Hollingshead, Yue Wang,
  Yujing Zhou
}
