% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint_functions.R
\name{p.value}
\alias{p.value}
\title{p-Value calculation for binary model}
\usage{
p.value(y, x, cpt)
}
\description{
DESC - Reminder to myself: x and y need to be sorted beforehand
}
\details{
DETAILS - Also, create function or text for separating data to find multiple
changepoints.
}
