% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint_functions.R
\name{preselect.poisson.calc}
\alias{preselect.poisson.calc}
\title{Preselected changepoint, Poisson parameter t.s.}
\usage{
preselect.poisson.calc(data, pre, minseglen)
}
\arguments{
\item{ts}{size m vector of timestamps associated with observed outcomes}

\item{theta}{single parameter, used with ts to inform transition matrix}

\item{obs}{n x m emission matrix of data observed at different timestamps (column) given different states (row)}
}
\value{
a vector representing the Viterbi path
}
\description{
This function implements a Viterbi algorithm, given m timestamps, single parameter theta, and n x m emission matrix.
}
\examples{
# TO BE DETERMINED

}
