% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunConsiderationSet.R
\name{RunConsiderationSet}
\alias{RunConsiderationSet}
\title{Consideration Set Model}
\usage{
RunConsiderationSet(YObservedInput, XConsidInput, XSelectInput, NumOpts,
  ASCConsid = 0, ASCSelect = 0, IncludedInput = 0 * YObservedInput + 1,
  meth = "L-BFGS-B", Tolerance = 1e-05)
}
\arguments{
\item{YObservedInput}{A vector of observed choices acting as response
variable}

\item{XConsidInput}{A vector or matrix of choices potential to be included in
consideration set}

\item{XSelectInput}{A vector or matrix of choices potential to be selected
from consideration set (i.e., must be among \code{XConsidInput})}

\item{NumOpts}{Number of options for consideration set}

\item{ASCConsid}{Binary flag for whether users want ASC (i.e., intercept) in
Consideration (i.e., choice). Default is 0.}

\item{ASCSelect}{Binary flag for whether users want ASC (i.e., intercept) in
Selection (i.e., choice). Default is 0.}

\item{IncludedInput}{A vector or matrix of which choices are available;
default is that every choice is available all the time.}

\item{meth}{Method of optimal function, default is method "L-BFGS-B". Other
methods available are "BFGS", "CG", "SANN", "Brent".}

\item{Tolerance}{Controls the convergence of the "L-BFGS-B" method.
Convergence occurs when the reduction in the objective is within this
factor of the machine tolerance. Default is a tolerance of about 1e-5.}
}
\value{
A list object containing likelihood for consideration set model,
  likelihood for logit model, and the coefficient for all coviriates in
  Consideration and Selection. If we set ASCConsid/ASCSelect to be 1, the
  output will include ASC in Consideration/Selection.
}
\description{
Constructs consideration set via binary logit model, then selects from
consideration set via multinomial logit model.
}
\examples{
\donttest{
data(CFS5-1)
attach(CFS5-1)
RunConsiderationSet(V1, cbind(V2,V3,V4), cbind(V2,V3,V4), 5, 1, 1,Tolerance=1e-6)

data(CFS6-1)
attach(CFS6-1)
RunConsiderationSet(V1, cbind(V2,V3,V4), cbind(V2,V3,V4), 6, 1, 1)

data(CFS5-2)
attach(CFS5-2)
RunConsiderationSet(V1, cbind(V2,V3,V4), cbind(V2,V3,V4), 5, 1, 1, IncludedInput=V5, meth="BFGS")

data(CFS6-2)
attach(CFS6-2)
RunConsiderationSet(V1, cbind(V2,V3,V4), cbind(V2,V3,V4), 6, 1, 1, IncludedInput=V5, meth="BFGS")
}

}
