% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se_outer}
\alias{se_outer}
\title{Standard errors of parameters and proportions (outer-version)}
\usage{
se_outer(out)
}
\arguments{
\item{out}{A list object obtained directly from \code{\link{LatentStage}}}
}
\value{
A list object with two attributes. \code{se.beta} is the
  outer-version SE of linear parameters. \code{se.lambda} is the
  outer-version SE of class proportions. Note in the
  \code{\link{LatentStage}} function that the latter is not provided.
}
\description{
Calculates outer-version standard errors for estimate of linear parameters
and class proportions.
}
