% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se_standard}
\alias{se_standard}
\title{Standard errors of parameters and proportions (standard version)}
\usage{
se_standard(out)
}
\arguments{
\item{out}{A list object obtained directly from \code{\link{LatentStage}}}
}
\value{
A list object with two attributes. \code{se.beta} is the standard SE
  of linear parameters. \code{se.lambda} is the standard SE of class
  proportions. Note in the \code{\link{LatentStage}} function that the latter
  is not provided.
}
\description{
Calculates standard errors for estimate of linear parameters and class
proportions.
}
