% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataFiles.R
\docType{data}
\name{SnackSet}
\alias{SnackSet}
\title{Snack Set}
\format{An object of class \code{data.frame} with 2376 rows and 14 columns.}
\usage{
SnackSet
}
\description{
In this dataset, there are 12 kinds of snacks for people to choose from.
Participants rated how much they liked each of the 12 snacks using an
11-point Likert scale (1 = "dislike very much" and 11 = "like very much"),
which gives us the covariate "Rate". Following that, participants ranked the
12 snacks, from their most favorite (rank = 1) to their least favorite (rank
= 12), which accounts for the covariate "Rank".
}
\keyword{datasets}
