% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataFiles.R
\docType{data}
\name{threestage}
\alias{threestage}
\title{Three Stage}
\format{An object of class \code{list} of length 4.}
\usage{
threestage
}
\description{
Structure of the data is single stage simulation, collapsing multistage model
into standard Discrete Heterogeneity Model.
}
\keyword{datasets}
